requirements.txt
streamlit
langchain
langchain-groq
langchain-classic
langchain-community
chromadb
pypdf
python-dotenv
sentence-transformers
